integration of rxn work
