+++ flux/Components/Home.js
@@ -1,65 +1,89 @@
-"use strict";
+'use strict';
 
 var React = require('react');
 
-module.exports = React.createClass({
+// COMPONENTS
+var HomeButton = React.createFactory(require('../Components/HomeButton.js'));
+
+// STORES
+var DisplayedItemStore = require('../Stores/displayedItemStore.js');
+
+// CONSTANTS
+var homeViewTypes = require('../Constants/constants.js').homeViewTypes;
+
+/*
+
+interface HomeProps{
+}
+interface HomeState{
+    view: string
+}
+
+*/
+
+function getStateFromStores() {
+    return {
+        view: DisplayedItemStore.getDisplayedView()
+    };
+}
+
+var Home = React.createClass({
 
     displayName: 'Home',
 
     getInitialState: function(){
-        return {};//getStateFromStores();
+        return getStateFromStores();
+    },
+
+    componentDidMount: function() {
+        // DisplayedItemStore.addChangeListener(this.onChange);
+    },
+
+    componentWillUnmount: function() {
+        // DisplayedItemStore.removeChangeListener(this.onChange);
     },
     
     render: function() {
-        //var self = this;
-        //var props = this.props;
-        //var state = this.state;
+        var state = this.state;
+
+        var searchBar = "Barre de recherche";
+
+        var view;
+
+        switch(state.view){
+
+            case homeViewTypes.SEARCH_RCS:
+                break;
+
+            // case homeViewTypes.RC_DETAILS:
+
+            // case homeViewTypes.SEARCH_ADS:
+
+            // case homeViewTypes.POST_AD:
+
+            default:
+                var buttons = [
+                    new HomeButton(),
+                    "Annonces",
+                    "Publier"
+                ];
+
+                view = React.DOM.div({},
+                    searchBar,
+                    buttons
+                );
+                break;
+
+        }
         
-        return React.DOM.div({className: 'home'}, [
-            
-            React.DOM.form({}, [
-                // mode switch
-                React.DOM.div({className: 'mode-switch'}, [
-                    React.DOM.label({}, [
-                        'Je me sépare',
-                        React.DOM.input({
-                            type: 'radio',
-                            id: 'give-away',
-                            name: 'give-search'
-                        })
-                    ]),
-                    React.DOM.label({}, [
-                        'Je récupère',
-                        React.DOM.input({
-                            type: 'radio',
-                            id: 'search-for',
-                            name: 'give-search'
-                        })
-                    ])
-                ]),
-                
-                React.DOM.label({}, [
-                    'Quoi ? ',
-                    React.DOM.input({
-                        type: 'text',
-                        id: 'what'
-                    })
-                ]),
-                
-                React.DOM.div({className: 'submit-group'}, [
-                    React.DOM.button({
-                        type: "submit"
-                    }, "Aller en déchèterie"),
-                    React.DOM.button({
-                        type: "submit"
-                    }, "Voir les personnes qui cherchent (21)"),
-                    React.DOM.button({
-                        type: "submit"
-                    }, "Déposer une annonce")
-                ])
-            ])
-            
-        ])
-        
+        return React.DOM.div({id: 'home'},
+            view
+        );
+    },
+
+    onChange: function() {
+        this.setState(getStateFromStores());
     }
 });
+
+module.exports = Home;
